% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cld.R
\name{as_pt_cld}
\alias{as_pt_cld}
\alias{as_pt_cld.pt_cld}
\alias{as_pt_cld.matrix}
\alias{as_pt_cld.data.frame}
\alias{as_pt_cld.default}
\title{Convert an object to a point cloud}
\usage{
as_pt_cld(x, ...)

\method{as_pt_cld}{pt_cld}(x, ...)

\method{as_pt_cld}{matrix}(x, ...)

\method{as_pt_cld}{data.frame}(x, ...)

\method{as_pt_cld}{default}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{pt_cld} object.
}
\description{
Generic function to convert various objects to a \code{pt_cld} point cloud object.
}
\examples{
df <- data.frame(x = 1:5, y = 6:10, z = 11:15)
cloud <- as_pt_cld(df)
}
