% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxdim.R
\name{boxdim}
\alias{boxdim}
\title{Compute the box dimension structural complexity index for a given point
cloud.}
\usage{
boxdim(
  cloud,
  threshold,
  vox_res = NULL,
  plot = TRUE,
  plot_title = NULL,
  warnings = TRUE
)
}
\arguments{
\item{cloud}{A \code{pt_cld} object (see \code{\link[=as_pt_cld]{as_pt_cld()}}).}

\item{threshold}{The lower resolution threshold until which the box dimension
algorithm iterates}

\item{vox_res}{Numeric of the resolution for voxelization. If left undefined
(default), the raw cloud will be analzed, which should only be done if the
cloud has been voxelized before independently.}

\item{plot}{Logical. If TRUE, a ggplot of the log-log regression is
included in the output. Defaults to TRUE.}

\item{warnings}{Logical controlling whether to display a warning when vox_res
is left undefined. Defaults to TRUE.}
}
\value{
If \code{plot = FALSE}, a numeric representing the box dimension.
If \code{plot = TRUE}, a named list with elements \code{boxdim} (the box dimension)
and \code{plot} (a ggplot object).
}
\description{
Calculate the box dimension as described in Seidel et al. (2018)
}
\examples{
sphere <- gen_sphere(1, 0.01, c(0, 0, 0))
boxdim(sphere, 0.1, 0.05)
}
