% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_to_mat.R
\name{cloud_to_mat}
\alias{cloud_to_mat}
\title{Convert tabular data to a point cloud matrix}
\usage{
cloud_to_mat(input, which = "xyz")
}
\arguments{
\item{input}{Any tabular data representing a point cloud.}

\item{which}{A string to decide which point columns the function returns.
Consists of the 1 to 3 characters "x", "y", or "z". Defaults to "xyz"
returning all columns.}
}
\value{
A \code{pt_cld} object (when \code{which = "xyz"}) or a matrix subset.
}
\description{
\code{cloud_to_mat()} is deprecated. Use \code{\link[=as_pt_cld]{as_pt_cld()}} or \code{\link[=pt_cld]{pt_cld()}} instead,
which return a proper \code{pt_cld} class object.
}
\examples{
sphere <- gen_sphere(1, 0.05)
sphere <- as.data.frame(sphere)
cloud <- as_pt_cld(sphere)
}
