% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_to_vector.R
\name{cloud_to_vector}
\alias{cloud_to_vector}
\title{Homogenize a given point cloud from any tabular format to a 3-column matrix.}
\usage{
cloud_to_vector(input, which = "xyz")
}
\arguments{
\item{input}{Any tabular data representing a point cloud.}

\item{which}{A string to decide which point columns the function returns.
Consists of the 1 to 3 characters "x", "y", or "z". Defaults to "xyz"
returning all columns.}
}
\value{
A matrix with 1 to 3 columns named "x", "y", "z" representing the
point cloud
}
\description{
Turns any tabular data (data.frame, matrix, tibble) into a 3-column matrix.
This format is used throughout this package. Additional columns other than
x, y, z will be removed.
Simple tabular data such as xyz point clouds are computed multiples faster
within algorithms in R when compared against data.frames and friends.
}
\examples{
sphere <- gen_sphere(1, 0.05)
sphere <- as.data.frame(sphere)
sphere$extra_data <- 1
cloud_to_vector(sphere, "xy")
}
