% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculate_c2c_dist}
\alias{.calculate_c2c_dist}
\title{Calculate cloud to cloud distances for every point of a query cloud to the
respective nearest points of the reference cloud with a kd-tree using the
\code{RANN::nn2()} function.}
\usage{
.calculate_c2c_dist(cloud_query, cloud_ref, max_dist = NULL)
}
\arguments{
\item{cloud_query}{The queried cloud as a matrix-like object, whose points
will be labelled with the distance to the nearest point of the reference
cloud.}

\item{cloud_ref}{The reference cloud as a matrix-like object, whose points
the query cloud will be compared against.}

\item{max_dist}{Numeric of an optional maximum distance for the
calculation. Points above this threshold will be discarded.}
}
\value{
Matrix of cloud_query with an additional column named "distance"
with the distances per point to cloud_ref
}
\description{
Calculate cloud to cloud distances for every point of a query cloud to the
respective nearest points of the reference cloud with a kd-tree using the
\code{RANN::nn2()} function.
}
\examples{
\dontrun{
sphere1 <- gen_sphere(1, 0.1, c(0, 0, 0))
sphere2 <- gen_sphere(1, 0.1, c(0, 0.5, 0))
.calculate_c2c_dist(sphere1, sphere2, max_dist = 0.5)
}
}
