% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_stratum_area}
\alias{.get_stratum_area}
\title{Calculate the x-y-area of a point cloud by extracting the concave or convex
hull and applying the shoelace algorithm.}
\usage{
.get_stratum_area(stratum, hull_type = c("concave", "convex"))
}
\arguments{
\item{stratum}{Point cloud as a matrix-like object}

\item{hull_type}{String of the type of hull algorithm. "concave" (default) or
"convex"}
}
\value{
Numeric of the area of the given point cloud in x-y dimension
}
\description{
Calculate the x-y-area of a point cloud by extracting the concave or convex
hull and applying the shoelace algorithm.
}
\examples{
\dontrun{
sphere_3d <- gen_sphere(1, 0.05)
.get_stratum_area(sphere_3d, "convex")
}
}
