% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hullify}
\alias{.hullify}
\title{Extract the concave or convex hull for a 2-dimensional point cloud.}
\usage{
.hullify(points, type = c("concave", "convex"), warnings = FALSE)
}
\arguments{
\item{points}{Matrix with the 2-D points to extract the hull from.}

\item{type}{String of the type of hull algorithm. "concave" (default) or
"convex".}

\item{warnings}{Logic whether to show warning when less than 3 points were
extracted.}
}
\value{
A Matrix with the points that make up the hull.
}
\description{
Extract the concave or convex hull for a 2-dimensional point cloud.
}
\examples{
\dontrun{
sphere_3d <- gen_sphere(1, 0.05)
.hullify(sphere_3d, "convex")
}
}
