% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pad_layers}
\alias{.pad_layers}
\title{Check if a list of point clouds representing layers is equal to a given
length and add empty layers if it is shorter.}
\usage{
.pad_layers(layers, max_height)
}
\arguments{
\item{layers}{A list of matrices representing point clouds.}

\item{max_height}{The length to compare the list's length against.}
}
\value{
A list of matrices representing point clouds with empty matrices
added if list was shorter than \code{max_height}.
}
\description{
Check if a list of point clouds representing layers is equal to a given
length and add empty layers if it is shorter.
}
\examples{
\dontrun{
list1 <- rep(list(gen_sphere(1, 0.5)), 3)
list2 <- rep(list(gen_sphere(1, 0.5)), 5)
max_length <- max(length(list1), length(list2))
list1 <- .pad_layers(list1, max_length)
list2 <- .pad_layers(list2, max_length)
}
}
