% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.round_n}
\alias{.round_n}
\title{Round a numeric variable to multiples of a given number.}
\usage{
.round_n(x, n)
}
\arguments{
\item{x}{Numeric to round.}

\item{n}{Numeric of multiple to round \code{x} to.}
}
\value{
Numeric with \code{x} rounded to multiple of \code{n}
}
\description{
Round a numeric variable to multiples of a given number.
}
\examples{
\dontrun{
num <- 4.54
.round_n(num, 0.1)
}
}
