% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.shoelace}
\alias{.shoelace}
\title{Apply the shoelace algorithm to a set of 2-D hull points to extract the area.}
\usage{
.shoelace(points)
}
\arguments{
\item{points}{Matrix with the hull points to compute the area from.}
}
\value{
Numeric of the area of the hull.
}
\description{
Apply the shoelace algorithm to a set of 2-D hull points to extract the area.
}
\examples{
\dontrun{
sphere_3d <- gen_sphere(1, 0.05)
hull <- .hullify(sphere_3d, "convex")
.shoelace(hull)
}
}
