% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stratify}
\alias{.stratify}
\title{Bin z-coordinates of a point cloud to a given resolution.}
\usage{
.stratify(cloud, strata_size)
}
\arguments{
\item{cloud}{Pointcloud as a 3 column matrix with column names "x", "y", "z".
Use cloud_to_vector() to get this format.}

\item{strata_size}{Numeric of the resolution of vertical (z-axis) strata.}
}
\value{
The point cloud as a three column matrix with column names
"x", "y", "z", with z-axis binned to the given resolution.
}
\description{
For CCI calculation vertical strata need to be created. This function does
this by rounding only the z-coordinate of clouds to multiples of the strata
resolution.
}
\examples{
\dontrun{
sphere <- gen_sphere(1, 0.01)
.stratify(sphere, 0.05)
}
}
