% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sphere.R
\name{gen_sphere}
\alias{gen_sphere}
\title{Generate an artifical voxelized point cloud of a sphere.}
\usage{
gen_sphere(r, res, center = c(0, 0, 0))
}
\arguments{
\item{r}{Numeric of the radius of the sphere.}

\item{res}{Numeric of the voxel resolution of the cloud.}

\item{center}{Three element vector (x, y, z) of the center point of the
cloud.}
}
\value{
A point cloud as a matrix with columns "x", "y", "z".
}
\description{
Generate an artifical voxelized point cloud of a sphere.
}
\examples{
# Generate 1 meter sphere with 5 cm voxel resolution at center (1,2,3)
sphere <- gen_sphere(1, 0.05, c(1, 2, 3))
}
