% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_cld.R
\name{is_pt_cld}
\alias{is_pt_cld}
\title{Check if an object is a pt_cld}
\usage{
is_pt_cld(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical: \code{TRUE} if \code{x} inherits from \code{pt_cld}, \code{FALSE} otherwise.
}
\description{
Check if an object is a pt_cld
}
\examples{
cloud <- gen_sphere(1, 0.5)
is_pt_cld(cloud)
}
