% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_laz.R
\name{save_to_laz}
\alias{save_to_laz}
\title{Save cloud to a .laz file}
\usage{
save_to_laz(cloud, file_path)
}
\arguments{
\item{cloud}{A \code{pt_cld} object. Use \code{\link[=as_pt_cld]{as_pt_cld()}} to convert.}

\item{file_path}{Character string specifying the output file path.}
}
\value{
None. Writes a LAZ file to disk.
}
\description{
Converts a cloud in matrix format to a .laz file using the rlas package.
}
\examples{
\dontrun{
  cloud <- gen_sphere(1, 0.3)
  save_to_laz(cloud, "cloud_out.laz")
}
}
