% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxdim.R
\name{boxdim}
\alias{boxdim}
\title{Compute the box dimension structural complexity index for a given point
cloud.}
\usage{
boxdim(cloud, threshold, vox_res = NULL, warnings = TRUE)
}
\arguments{
\item{cloud}{Point cloud as a matrix-like object.}

\item{threshold}{The lower resolution threshold until which the box dimension
algorithm iterates}

\item{vox_res}{Numeric of the resolution for voxelization. If left undefined
(default), the raw cloud will be analzed, which should only be done if the
cloud has been voxelized before independently.}

\item{warnings}{Logical controlling whether to display a warning when vox_res
is left undefined. Defaults to TRUE.}
}
\value{
A numeric representing the box dimension.
}
\description{
Calculate the box dimension as described in Seidel et al. (2018)
}
\examples{
sphere <- gen_sphere(1, 0.01, c(0, 0, 0))
boxdim(sphere, 0.1, 0.05)
}
