% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi.R
\name{coi}
\alias{coi}
\title{Compute the crown overlap index (COI) for two raw single tree point clouds.}
\usage{
coi(cloud_i, cloud_j, d_max, vox_res = NULL, warnings = TRUE)
}
\arguments{
\item{cloud_i}{First single tree point cloud as a \code{pt_cld} object
(see \code{\link[=as_pt_cld]{as_pt_cld()}}).}

\item{cloud_j}{Second single tree point cloud as a \code{pt_cld} object
(see \code{\link[=as_pt_cld]{as_pt_cld()}}).}

\item{d_max}{Numeric of the maximum distance of interaction. Decides how far
away points of one tree can be from the other before they are excluded from
the interaction cloud. Set to 0 if only exact point or voxel matches should
contribute.}

\item{vox_res}{Numeric of the resolution for voxelization during
pre-processing. If left undefined (default), the raw clouds will be analzed,
which should only be done if the clouds have been voxelized before
independently.}

\item{warnings}{Logical controlling whether to display a warning when vox_res
is left undefined. Defaults to TRUE.}
}
\value{
A numeric representing the COI.
}
\description{
Calculate the COI as described in Täuber et al. (in prep.) All-in-one
function handling voxelization, interaction extraction, and calculation of
COI.
}
\details{
Takes two raw point clouds representing interacting trees
and voxelizes them to a given common point density. Then the interaction
(overlap) of the clouds is extracted according to the maximum distance
parameter. Finally from this interaction cloud the COI is computed.
This is a convenience function doing all preprocessing for COI calculation in
one function. You may want to inspect the interaction clouds before COI
computation or save them as seperate files. For this the low-level function
\code{compute_coi()} is available.
}
\examples{
sphere1 <- gen_sphere(1, 0.01, c(0, 0, 0))
sphere2 <- gen_sphere(1, 0.01, c(0, 0.5, 0))
coi(sphere1, sphere2, 0.3, 0.05)
}
