% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enl.R
\name{enl}
\alias{enl}
\title{Compute the Effective Number of Layers (ENL) for a voxelized point cloud.}
\usage{
enl(cloud, voxel_res, layer_thickness = 1, plot = TRUE, plot_title = NULL)
}
\arguments{
\item{cloud}{A \code{pt_cld} object. Use \code{\link[=as_pt_cld]{as_pt_cld()}} to convert.}

\item{voxel_res}{Numeric voxel resolution for voxelization.}

\item{layer_thickness}{Numeric thickness of each vertical layer (default 1).}

\item{plot}{Logical, whether to plot the vertical voxel distribution (default FALSE).}

\item{plot_title}{Optional character string for the plot title.}
}
\value{
If plot=FALSE, returns a list (ENL0, ENL1, ENL2). If plot=TRUE, returns list plus ggplot object.
}
\description{
Calculates ENL₀, ENL₁, ENL₂ as vertical structural complexity indices.
Optionally plots the vertical voxel distribution.
}
\examples{
sphere <- gen_sphere(1, 0.01)
enl_stats <- enl(sphere, voxel_res = 0.05, layer_thickness = 1, plot = TRUE)
}
