% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_interaction.R
\name{extract_interaction}
\alias{extract_interaction}
\title{Extract interaction distances between two point clouds.}
\usage{
extract_interaction(cloud_i, cloud_j, d_max, returns = c("distances", "cloud"))
}
\arguments{
\item{cloud_i}{First point cloud as a matrix-like object.}

\item{cloud_j}{Second point cloud as a matrix-like object.}

\item{d_max}{Numeric of the maximum distance to consider for nearest neighbor
search.}

\item{returns}{Character, either "distances" (default) to return a vector of
distances (to use in \code{calculate_coi()}), or "cloud" to return the combined
cloud with distances (useful when extracted interaction point cloud needs to
be inspected visually).}
}
\value{
Either a numeric vector of distances or a point cloud as a matrix
with columns "x", "y", "z", "distance".
}
\description{
Calculates the nearest neighbor distances between two point clouds in both
directions, returning either a numeric vector of the distances used to
compute the coi with \code{coi()} or the interaction point cloud with distances as
a fourth column.
}
\examples{
sphere1 <- gen_sphere(1, 0.05, c(0, 0, 0))
sphere2 <- gen_sphere(1, 0.05, c(0, 0.5, 0))
spheres_dists <- extract_interaction(sphere1, sphere2, d_max = 0.5)
}
