% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{voxelize}
\alias{voxelize}
\title{Voxelize matrix-like point cloud objects.}
\usage{
voxelize(cloud, res)
}
\arguments{
\item{cloud}{A \code{pt_cld} object (see \code{\link[=as_pt_cld]{as_pt_cld()}}).}

\item{res}{Numeric of the voxel resolution to apply on the point cloud.}
}
\value{
A \code{pt_cld} object with the voxelized point cloud.
}
\description{
This is a simple round-to-multiple function which can be used to voxelize
point clouds.
}
\examples{
sphere <- gen_sphere(1, 0.01)
sphere_v <- voxelize(sphere, 0.05)
}
