% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{voxelize}
\alias{voxelize}
\title{Voxelize matrix-like point cloud objects.}
\usage{
voxelize(cloud, res)
}
\arguments{
\item{cloud}{The point cloud to voxelize as a matrix-like object.}

\item{res}{Numeric of the voxel resolution to apply on the point cloud.}
}
\value{
The voxelized point cloud as a matrix with columns "x", "y", "z".
}
\description{
This is a simple round-to-multiple function which can be used to voxelize
point clouds.
}
\examples{
sphere <- gen_sphere(1, 0.01)
voxelize(sphere, 0.05)
}
